third_party/blink/web_tests/platform/mac-mac11-arm64/external/wpt/IndexedDB/idbobjectstore-rename-abort-expected.txt

This is a testharness.js-based test.
FAIL IndexedDB object store rename in aborted transaction assert_equals: Expected success event, but got error event instead expected "success" but got "error"
PASS IndexedDB object store creation and rename in an aborted transaction
Harness: the test ran to completion.
